fix(section): toggleState not being executed onClick
In this commit, we have resolved an issue where `toggleState` function in our `Section` component wasn't executed when an element was clicked. This has been achieved by changing the `toggleState` function return type to Promise and modifying its implementation, along with some changes in the `OptionButton` and `Section.test` files. The test cases have also been altered with the introduction of a new assertion to check if `setStorage` was called upon clicking the setting. This change ensures that upon interaction, the state of the settings updates as required.